% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_bxr_v4.R
\name{openBXR}
\alias{openBXR}
\title{Open and validate BXR file}
\usage{
openBXR(file)
}
\arguments{
\item{file}{Character. Path to BXR file (.bxr extension)}
}
\value{
H5File object from hdf5r package for reading BXR data
}
\description{
Opens a BXR file
}
\examples{
\dontrun{
# Open BXR file
h5 <- openBXR("experiment.bxr")

# Always close when done
h5$close()
}
}
