% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction.R
\name{eventBased}
\alias{eventBased}
\title{Data Extraction and Parsing Functions}
\usage{
eventBased(h5, well_id, target_chunks)
}
\arguments{
\item{h5}{H5File object}

\item{well_id}{Character. Well identifier (e.g., "Well_A1")}

\item{target_chunks}{Integer vector. Chunk indices to extract}
}
\value{
List containing binary_chunk, stored_channels, and data_range
}
\description{
Functions for extracting and parsing EventsBasedSparseRaw data
}
\author{
hdf5MEA package
Extract EventsBasedSparseRaw binary data

Extracts binary data from EventsBasedSparseRaw dataset for specified chunks.
}
