% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_brw_v4.R
\name{openBRW}
\alias{openBRW}
\title{BRW v4 File I/O Functions}
\usage{
openBRW(file)
}
\arguments{
\item{file}{Character. Path to BRW file}
}
\value{
H5File object from hdf5r package
}
\description{
Functions for opening BRW v4 files and processing time/chunk information
}
\examples{
\dontrun{
h5 <- openBRW("data.brw")
h5$close()
}
}
\author{
hdf5MEA package
Open and validate BRW file

Opens a BRW file and validates it's a proper HDF5 file structure.
}
